requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.2.5
django-cors-headers==3.7.0
django-seed==0.3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
eunjeon==0.4.0
Faker==9.5.0
flex==6.14.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonpointer==2.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.1
packaging==21.0
Pillow==8.3.1
pyparsing==2.4.7
pytagcloud==0.3.5
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
strict-rfc3339==0.7
text-unidecode==1.3
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.7
validate-email==1.3
wordcloud==1.8.1
